requirements.txt
peewee>=3.14.0
aiohttp>=3.7.3
zhconv>=1.4.1
requests>=2.25.1
nonebot-adapter-cqhttp>=2.0.0a11.post2
nonebot2>=2.0.0a11
numpy>=1.20.1
Pillow>=8.1.0
nonebot-plugin-apscheduler>=0.1.2
lxml>=4.6.2
beautifulsoup4>=4.9.3
feedparser>=6.0.2
brotli>=1.0.9
psutil>=5.8.0
opencv-python>=4.5.1.48
aiofiles>=0.6.0
pyjwt>=2.1.0
sqlitedict>=1.7.0
aioaria2
